Rename logger variable
